Move models to own files. Apply OCA conventions
