+++ vault/models/__init__.py
@@ -4,9 +4,12 @@
 from . import (
     abstract_vault,
     res_users,
+    res_users_key,
     vault,
     vault_entry,
     vault_field,
     vault_inbox,
+    vault_log,
     vault_right,
+    vault_tag,
 )
