+++ vault/models/abstract_vault.py
@@ -3,48 +3,10 @@
 
 import logging
 
-from odoo import _, api, fields, models
+from odoo import _, api, models
 from odoo.exceptions import AccessError
 
 _logger = logging.getLogger(__name__)
-
-
-class VaultLog(models.Model):
-    _name = "vault.log"
-    _description = _("Log entry of a vault")
-    _order = "create_date DESC"
-    _rec_name = "message"
-
-    def _get_log_state(self):
-        return [
-            ("info", _("Information")),
-            ("warn", _("Warning")),
-            ("error", _("Error")),
-        ]
-
-    vault_id = fields.Many2one(
-        "vault",
-        "Vault",
-        ondelete="cascade",
-        required=True,
-        readonly=True,
-    )
-    entry_id = fields.Many2one(
-        "vault.entry",
-        "Entry",
-        ondelete="cascade",
-        readonly=True,
-    )
-    user_id = fields.Many2one("res.users", "User", required=True, readonly=True)
-    state = fields.Selection(_get_log_state, readonly=True)
-    message = fields.Char(readonly=True, required=True)
-
-    @api.model
-    def create(self, values):
-        res = super().create(values)
-        if not self.env.context.get("skip_log", False):
-            _logger.info("Vault log: %s", res.message)
-        return res
 
 
 class AbstractVault(models.AbstractModel):
